chore: rearrange simple message queue code
